frontend/requirements.txt
streamlit
requests
python-dotenv
pycountry
